+++ pages/dashboards/tasks/index.tsx
@@ -17,7 +17,7 @@ import SourceTable from "@/content/Dashboards/Tasks/SourceTable";
 
 import { getDashboardData, getReviewsData } from "@/services";
 import DataContext from "@/contexts/DataContext";
-import SocialPages from "@/content/Dashboards/Tasks/AddSocialPages";
+import SocialPages from "@/content/Dashboards/Tasks/SocialPages";
 import CustomAlert from "@/components/CustomAlert";
 
 const TabsContainerWrapper = styled(Box)(
