+++ src/services/socialMediaLink/index.ts
@@ -1,6 +1,6 @@
 import api from "../api/index";
 
-export const getClientSocialMedia = async (clientid: string, title?: string) => {
+export const getClientSocialMediaLink = async (clientid: string, title?: string) => {
   return await new Promise((resolve, reject) => {
     api
       .get("/getClientSocialmedialink", {
@@ -20,11 +20,29 @@ export const getClientSocialMedia = async (clientid: string, title?: string) => 
   });
 };
 
-export const createClientSocialmedialink = async (params) => {
+export const createClientSocialMediaLink = async (params) => {
   return await new Promise((resolve, reject) => {
     api
       .post("/createClientSocialmedialink", params)
       .then((response) => {
+        console.log(response);
+        if (response?.data?.status === "success") resolve("success");
+        else reject("Something went wrong");
+      })
+      .catch((error) => {
+        console.log(error);
+        if (error?.message) reject(error.message);
+        else reject("Something went wrong");
+      });
+  });
+};
+
+export const deleteClientSocialMediaLink = async (id: number) => {
+  return await new Promise((resolve, reject) => {
+    api
+      .post("/deleteClientSocialmedialink", { id })
+      .then((response) => {
+        console.log(response);
         if (response?.data?.status === "success") resolve("success");
         else reject("Something went wrong");
       })
